Viet anh Update new project

+++ migrations/20190506162055_Admin.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
         table.string('fullname', 50).notNullable();
         table.integer('age', 10).notNullable();
         table.text('address', 50).notNullable();
-        table.tinyInteger('gender').notNullable();
+        table.integer('gender').notNullable();
         table.integer('phone', 20).notNullable();
         table.string('username', 50).notNullable();
         table.string('password', 50).notNullable();
