+++ migrations/20190506162312_Bills.js
@@ -6,7 +6,7 @@ exports.up = function(knex, Promise) {
         table.foreign('id_customer').references('id').inTable('Customer');
         table.dateTime('date_order').notNullable();
         table.integer('total').notNullable();
-        table.tinyInteger('payment_status').notNullable();
+        table.integer('payment_status').notNullable();
         table.text('note', 100).notNullable();
     });
 };
